demos/yolov3_huaweiShip/requirements.txt
albumentations==1.1.0
backports.lzma==0.0.14
cycler==0.11.0
decorator==4.4.2
fonttools==4.28.5
imageio==2.13.5
joblib==1.1.0
kiwisolver==1.3.1
matplotlib==3.3.4
networkx==2.5.1
numpy==1.19.5
nvidia-dali-cuda110==1.9.0
opencv-python==4.5.4.60
opencv-python-headless==4.5.5.62
packaging==21.3
Pillow==8.4.0
pyparsing==3.0.6
python-dateutil==2.8.2
PyWavelets==1.1.1
PyYAML==6.0
qudida==0.0.4
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
six==1.16.0
threadpoolctl==3.0.0
tifffile==2020.9.3
torch==1.10.1+cu113
torchaudio==0.10.1+cu113
torchvision==0.11.2+cu113
tqdm==4.62.3
typing_extensions==4.0.1
